requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.1
sqlparse==0.5.0
